wheat_disease_classification/requirements.txt
torch
torchvision
scikit-learn
tqdm
Pillow
